Refactor metrics handling in ApproveBlockProtocol
- Introduced a new `Metrics` trait for better metrics management.
- Implemented `MetricsTestImpl` for isolated testing of metrics functionality.
- Updated `ApproveBlockProtocolImpl` to include error handling when incrementing metrics counters.
- Removed the previous metrics trait implementation from `approve_block_protocol.rs` to streamline the codebase.
