+++ shared/src/rust/shared/mod.rs
@@ -2,5 +2,7 @@ pub mod compression;
 pub mod f1r3fly_event;
 pub mod f1r3fly_events;
 pub mod list_ops;
+pub mod metrics;
+pub mod metrics_test;
 pub mod printer;
 pub mod string_ops;
